# src/app/api/docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This is Almond Backend API server documentation.For this sample, you can use the api key `special-key` to test the authorization     filters.",
		"version": "1.0.0",
		"title": "Almond Backend API",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "email@email.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "localhost:8080",
	"basePath": "/api",
...
	"schemes": ["http"],
	"paths": {
...
		"/pump": {
			"patch": {
				"tags": ["device"],
				"summary": "Edit pump override on Almond Hydroponics backend API",
				"description": "This Endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Device"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			},
			"get": {
				"tags": ["roles"],
				"summary": "Get user roles for Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
...
		"/devices": {
			"get": {
				"tags": ["device"],
				"summary": "Get all devices on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			},
			"post": {
				"tags": ["device"],
				"summary": "Add a new device on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/DeviceId"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
		"/devices/{id}": {
			"patch": {
				"tags": ["device"],
				"summary": "Edit a device on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/DeviceId"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			},
			"delete": {
				"tags": ["device"],
				"summary": "Delete a device by id on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "device Id",
						"required": true
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
		"/my-device": {
			"post": {
				"tags": ["device"],
				"summary": "Add device verification on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/DeviceId"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
		"/active-device": {
			"patch": {
				"tags": ["device"],
				"summary": "Update active device on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/DeviceId"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
...
		"/link_account/confirm_token_google": {
			"post": {
				"tags": ["linkAccounts"],
				"summary": "Link Accounts on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/LinkAccount"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		},
		"/link_account/google": {
			"post": {
				"tags": ["linkAccounts"],
				"summary": "Link Accounts google on Almond Hydroponics backend API",
				"description": "This endpoint Description",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "",
						"required": true,
						"schema": {
							"$ref": "#/definitions/LinkAccount"
						}
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					},
					"405": {
						"description": "Method Not Allowed"
					}
				},
				"security": [
					{
						"api_key": []
					}
				]
			}
		}
	},
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
	"definitions": {
...
		"Device": {
			"type": "object",
			"required": ["enabled", "device"],
			"properties": {
				"enabled": {
					"type": "string",
					"example": "Boolean (true/false)"
				},
				"device": {
					"type": "string",
					"example": "Device Number/Name"
				}
			}
		},
		"DeviceId": {
			"type": "object",
			"required": ["enabled", "device"],
			"properties": {
				"id": {
					"type": "string",
					"example": "Device ID"
				}
			}
		},
...
		"LinkAccount": {
			"type": "object",
			"required": ["token"],
			"properties": {
				"token": {
					"type": "string",
					"example": "token"
				},
				"email": {
					"type": "email",
					"example": "token"
				}
			}
		},
		"ApiResponse": {
			"type": "object",
			"properties": {
				"code": {
					"type": "integer",
					"format": "int32"
				},
				"type": {
					"type": "string"
				},
				"message": {
					"type": "string"
				}
			}
		}
	},
...
}